% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histDataDS.R
\name{histDataDS}
\alias{histDataDS}
\title{Generate histogram data for plotting}
\usage{
histDataDS(x, breaks = 10)
}
\arguments{
\item{x}{a numeric vector}

\item{breaks}{number of breaks for histogram}
}
\value{
a list containing histogram data and summary stats
}
\description{
Returns aggregated histogram data that can be safely shared
}
\author{
Your Name
}
